LPS-66926 Move calendar data handlers to the proper packages
